--- docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks/variables/CHATS_LIST_MOCK.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: CHATS\_LIST\_MOCK
-
-> `const` **CHATS\_LIST\_MOCK**: `InterfaceGQLMock`[]
-
-Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:562](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L562)
