+++ docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks/variables/CHAT_BY_ID_QUERY_MOCK.md
@@ -6,4 +6,4 @@
 
 > `const` **CHAT\_BY\_ID\_QUERY\_MOCK**: `InterfaceGQLMock`[]
 
-Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:297](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L297)
+Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:294](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L294)
