--- docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks/variables/CREATE_CHAT_MUTATION_MOCK.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: CREATE\_CHAT\_MUTATION\_MOCK
-
-> `const` **CREATE\_CHAT\_MUTATION\_MOCK**: `InterfaceGQLMock`[]
-
-Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:730](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L730)
