--- docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks/variables/GROUP_CHAT_BY_ID_QUERY_MOCK.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: GROUP\_CHAT\_BY\_ID\_QUERY\_MOCK
-
-> `const` **GROUP\_CHAT\_BY\_ID\_QUERY\_MOCK**: `InterfaceGQLMock`[]
-
-Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:720](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L720)
