+++ docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks/variables/GROUP_CHAT_BY_USER_ID_QUERY_MOCK.md
@@ -6,4 +6,4 @@
 
 > `const` **GROUP\_CHAT\_BY\_USER\_ID\_QUERY\_MOCK**: `InterfaceGQLMock`[]
 
-Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:418](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L418)
+Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:415](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L415)
