--- docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks/variables/MARK_CHAT_MESSAGES_AS_READ_MOCK.md
@@ -1,9 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Variable: MARK\_CHAT\_MESSAGES\_AS\_READ\_MOCK
-
-> `const` **MARK\_CHAT\_MESSAGES\_AS\_READ\_MOCK**: `InterfaceGQLMock`[]
-
-Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:805](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L805)
