+++ docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks/variables/MESSAGE_SENT_TO_CHAT_MOCK.md
@@ -6,4 +6,4 @@
 
 > `const` **MESSAGE\_SENT\_TO\_CHAT\_MOCK**: `InterfaceGQLMock`[]
 
-Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:279](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L279)
+Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:276](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L276)
