+++ docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks/variables/UserConnectionListMock.md
@@ -6,4 +6,4 @@
 
 > `const` **UserConnectionListMock**: `InterfaceGQLMock`[]
 
-Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:199](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L199)
+Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts:196](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks.ts#L196)
