+++ docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2/variables/CHATS_LIST_MOCK.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: CHATS\_LIST\_MOCK
+
+> `const` **CHATS\_LIST\_MOCK**: `InterfaceGQLMock`[]
+
+Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2.ts:161](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2.ts#L161)
