+++ docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2/variables/CREATE_CHAT_MUTATION_MOCK.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: CREATE\_CHAT\_MUTATION\_MOCK
+
+> `const` **CREATE\_CHAT\_MUTATION\_MOCK**: `InterfaceGQLMock`[]
+
+Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2.ts:329](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2.ts#L329)
