+++ docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2/variables/GROUP_CHAT_BY_ID_QUERY_MOCK.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: GROUP\_CHAT\_BY\_ID\_QUERY\_MOCK
+
+> `const` **GROUP\_CHAT\_BY\_ID\_QUERY\_MOCK**: `InterfaceGQLMock`[]
+
+Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2.ts:319](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2.ts#L319)
