+++ docs/docs/auto-docs/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2/variables/MARK_CHAT_MESSAGES_AS_READ_MOCK.md
@@ -1,0 +1,9 @@
+[Admin Docs](/)
+
+***
+
+# Variable: MARK\_CHAT\_MESSAGES\_AS\_READ\_MOCK
+
+> `const` **MARK\_CHAT\_MESSAGES\_AS\_READ\_MOCK**: `InterfaceGQLMock`[]
+
+Defined in: [src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2.ts:404](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/UserPortal/CreateDirectChat/CreateDirectChatMocks2.ts#L404)
